chore: split minimap functions
